[PATCH] Display SPI transfer agreement in common code
[ This is an updated version of
  http://marc.theaimsgroup.com/?l=linux-scsi&m=110918302718178&w=2
  Relative to that patch, it prints the period in a nicer way; exactly
  the same way as the strings in the original version of this file. ]

Introduce a generic SPI function to print the negotiated transfer
agreement.  It's based on the implementation in sym53c8xx_2/sym_misc.c.
I have changes in my tree to make 53c700, ncr53c8xx and sym2 use it.
Other drivers need to be converted to use the SPI transport layer first.

In order to calculate the speed I needed to be able to convert from
period factor to period in picoseconds.  That required changing
(show|store)_spi_transport_period to work in picoseconds rather than
a string.

Signed-off-by: Matthew Wilcox <matthew@wil.cx>
Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>