[PATCH] USB: Make use_both_schemes=y the default
Enough people are experiencing problems with the new device initialization
scheme that it seems like a good idea to make the default behavior be to
try the old scheme when the new one fails.  That's what this patch does,
simply by changing the initial value of the use_both_schemes module
parameter.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>