[PATCH] USB: add usbmon, a USB monitoring framework
This patch adds so-called "usbmon", or USB monitoring framework, similar
to what tcpdump provides for Ethernet. This is an initial version, but
it should be safe and useful. It adds an overhead of an if () statement
into submission and giveback paths even when not monitoring, but this
was deemed a lesser evil than stealth manipulation of function pointers.

The patch makes two changes to hcd.c which make usbmon more useful:
 - Change the way we determine that DMA should not be mapped for root
   hubs, so that usbmon knows easily when it's safe to capture data.
 - Return exports of usb_bus_list and usb_bus_list_lock for those who
   wish to build usbmon as a module.

This version of the patch changes #define to inlines for hooks and
drops extra mod_ops.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>