[PATCH] USB: don't power down net2280 on suspend
This creates a module parameter "enable_suspend", and changes the default
behavior to avoid the need to power cycle machines with net2280 cards to
recover from some common developer actions (like rmmod/modprobe).

When set to 1, the net2280, for better or worse, will enter low-power
mode when the USB host requests it.  This is fine except in situations
involving a power-cycling host or rmmoding the gadget driver while
connected to a suspended (or disconnected!) port.

When set to 0 (now the default), the driver will report suspend requests
to the gadget driver, but will not place the NET2280 into low-power mode.
This works fine in all situations except bus-powered devices (for which
a PCI+Linux-based NET2280 generally doesn't make much sense anyway)

It also (finally...) updates NetChip references to refer to PLX; and
updates the dates on the driver.

From: Alex Sanks <ASanks@plxtech.com>
Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>