[PATCH] USB: Fix race in URB submission vs. endpoint-disable
This patch fixes a race between URB submission and endpoint-disable.  (I
don't know that it's actually possible to trigger the race, but it won't
hurt to move the appropriate test inside the region protected by the
spinlock.)  In addition, an out-of-date comment referring to timeouts is
removed.

The patch also changes the error code returned when an URB is submitted
for a nonexistent or disabled endpoint to -ENOENT and documents the error
code (which was already in use by usbfs) in
Documentation/usb/error-codes.txt.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>