[PATCH] sysdev: remove the rwsem usage from this subsystem.
If further finer grained locking is needed, we can add a lock to the sysdev_class to
lock the class drivers list.  But if you do that, remember the global list also is still
there and needs to be protected.  That's why I went with a simple lock for everything.

Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>