[PATCH] x86: disable MSI for AMD-8131
The AMD-8131 I/O APIC (device id 1022:7450/7451) does not support message
signalled interrupts.  Thus, if a device driver attempts to enable msi, it
will suceed, but interrupts are not actually delivered to the cpu.  The
Nforce chipsets do not seem to have this limitation.  AMD confirmed that
MSI mode is unsupported with this APIC.

The following patch adds a flag to pci quirks to detect this and disable
msi.

Signed-off-by: Michael S. Tsirkin <mst@mellano.co.il>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>